fix: various fixes after merges and from previous branch state
